+++ src/Checkout/Payment/Service/OrderExecuteService.php
@@ -63,7 +63,7 @@ class OrderExecuteService
             return $this->doPayPalRequest($paypalOrder, $salesChannelId, $partnerAttributionId, $transactionId, $context);
         } catch (PayPalApiException $e) {
             if ($e->getStatusCode() !== Response::HTTP_UNPROCESSABLE_ENTITY
-                || ($e->getIssue() !== PayPalApiException::ERROR_CODE_DUPLICATE_INVOICE_ID)) {
+                || !$e->is(PayPalApiException::ERROR_CODE_DUPLICATE_INVOICE_ID)) {
                 throw $e;
             }