+++ src/Resources/app/administration/src/module/swag-paypal-disputes/page/swag-paypal-disputes-list/index.ts
@@ -23,7 +23,7 @@ export default Shopware.Component.wrapComponentConfig({
     ],
 
     mixins: [
-        Shopware.Mixin.getByName('notification'),
+        Shopware.Mixin.getByName('swag-paypal-notification'),
     ],
 
     data(): {
@@ -173,11 +173,12 @@ export default Shopware.Component.wrapComponentConfig({
         }, 850),
 
         handleError(errorResponse: PayPal.ServiceError) {
-            const errorDetail = errorResponse.response?.data.errors?.[0]?.detail;
-            this.createNotificationError({
-                message: `${this.$tc('swag-paypal-disputes.list.errorTitle')}: ${errorDetail}`,
-                autoClose: false,
-            });
+            if (errorResponse.response?.data.errors?.[0]?.code === DISPUTE_AUTH_ERROR) {
+                this.notAuthorized = true;
+            } else {
+                this.createNotificationFromError({ errorResponse, title: 'swag-paypal-disputes.list.errorTitle' });
+            }
+
             this.isLoading = false;
         },