+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-checkout/swag-paypal-checkout.spec.js
@@ -4,7 +4,6 @@ import 'src/app/component/base/sw-container';
 import 'src/app/component/base/sw-icon';
 import 'src/app/component/base/sw-alert';
 import 'src/app/mixin/notification.mixin';
-import 'SwagPayPal/mixin/swag-paypal-credentials-loader.mixin';
 import 'SwagPayPal/module/swag-paypal/components/swag-paypal-checkout';
 
 Shopware.Component.register('swag-paypal-checkout', () => import('.'));