+++ src/RestApi/Client/AbstractClient.php
@@ -9,14 +9,22 @@ namespace Swag\PayPal\RestApi\Client;
 
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
+use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
 use Shopware\Core\Framework\Log\Package;
 use Swag\PayPal\RestApi\Exception\PayPalApiException;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 #[Package('checkout')]
 abstract class AbstractClient
 {
+    protected const HEADER_WHITELIST = [
+        'Paypal-Debug-Id',
+        'PayPal-Request-Id',
+        'Date',
+    ];
+
     protected ClientInterface $client;
 
     protected LoggerInterface $logger;
@@ -127,21 +135,26 @@ abstract class AbstractClient
         $exceptionResponse = $requestException->getResponse();
 
         if ($exceptionResponse === null) {
-            $this->logger->error($exceptionMessage, [$data]);
+            $this->logger->error($exceptionMessage, ['data' => $data]);
 
             return new PayPalApiException('General Error', $exceptionMessage);
         }
 
         $content = $exceptionResponse->getBody()->getContents();
         $error = \json_decode($content, true) ?: [];
+        $issue = null;
         if (\array_key_exists('error', $error) && \array_key_exists('error_description', $error)) {
             $this->logger->error($exceptionMessage, [
                 'error' => $error,
-                'headers' => $exceptionResponse->getHeaders(),
+                'headers' => $this->extractHeaders($exceptionResponse),
                 'data' => $data,
             ]);
 
-            return new PayPalApiException($error['error'], $error['error_description'], (int) $requestException->getCode());
+            if ($exceptionResponse->getStatusCode() === Response::HTTP_UNAUTHORIZED) {
+                $issue = PayPalApiException::ERROR_CODE_INVALID_CREDENTIALS;
+            }
+
+            return new PayPalApiException($error['error'], $error['error_description'], $exceptionResponse->getStatusCode(), $issue);
         }
 
         if (\is_array($error['errors'] ?? null)) {
@@ -149,7 +162,6 @@ abstract class AbstractClient
         }
 
         $message = $error['message'] ?? $content;
-        $issue = null;
 
         if (isset($error['details'])) {
             $message .= ' ';
@@ -169,10 +181,18 @@ abstract class AbstractClient
 
         $this->logger->error(\sprintf('%s %s', $exceptionMessage, $message), [
             'error' => $error,
-            'headers' => $exceptionResponse->getHeaders(),
+            'headers' => $this->extractHeaders($exceptionResponse),
             'data' => $data,
         ]);
 
-        return new PayPalApiException($error['name'] ?? 'UNCLASSIFIED_ERROR', $message, (int) $requestException->getCode(), $issue);
+        return new PayPalApiException($error['name'] ?? 'UNCLASSIFIED_ERROR', $message, $exceptionResponse->getStatusCode(), $issue);
+    }
+
+    private function extractHeaders(ResponseInterface $response): array
+    {
+        return \array_combine(
+            self::HEADER_WHITELIST,
+            \array_map(static fn (string $name) => $response->getHeader($name), self::HEADER_WHITELIST),
+        );
     }
 }